+++ include/process.h
@@ -12,7 +12,7 @@ class Process {
   int Pid() const;                               // TODO: See src/process.cpp
   std::string User() const;                      // TODO: See src/process.cpp
   std::string Command();                   // TODO: See src/process.cpp
-  long CpuUtilization () const;                  // TODO: See src/process.cpp
+  float CpuUtilization () const;                  // TODO: See src/process.cpp
   std::string Ram() const;                       // TODO: See src/process.cpp
   long UpTime() const;                       // TODO: See src/process.cpp
   bool operator<(Process const& a) const;  // TODO: See src/process.cpp
